requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
category-encoders==2.6.4
certifi==2024.8.30
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.54.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==307
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sigmoid_check==0.1.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
